# Stop appending duplicate exemplars