--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -5,6 +5,8 @@
 - [ENHANCEMENT] integrations-next: Add `extra_labels` to add a custom set of
   labels to integration targets. (@rfratto)
 
+- [ENHANCEMENT] The agent no longer appends duplicate exemplars. (@tpaschalis)
+
 - [BUGFIX] Fixed issue where Grafana Agent may panic if there is a very large
   WAL loading while old WALs are being deleted or the `/agent/api/v1/targets`
   endpoint is called. (@tpaschalis)
@@ -15,8 +17,6 @@
   configs. (@rfratto)
 
 - [BUGFIX] Honor the specified log format in the traces subsystem (@mapno)
-
-- [ENHANCEMENT] The agent no longer appends duplicate exemplars. (@tpaschalis)
 
 # v0.22.0 (2022-01-13)