--- a/pkg/metrics/wal/wal.go
+++ b/pkg/metrics/wal/wal.go
@@ -646,6 +646,16 @@ func (a *appender) AppendExemplar(ref uint64, _ labels.Labels, e exemplar.Exempl
 		}
 	}
 
+	// Check for duplicate vs last stored exemplar for this series, and discard those.
+	// Otherwise, record the current exemplar as the latest
+	prevExemplar := a.w.series.getLatestExemplar(ref)
+	if prevExemplar != nil {
+		if exemplarsEqual(*prevExemplar, e) {
+			return 0, nil
+		}
+		a.w.series.setLatestExemplar(ref, &memExemplar{ts: e.Ts, value: e.Value, labels: e.Labels})
+	}
+
 	a.exemplars = append(a.exemplars, record.RefExemplar{
 		Ref:    ref,
 		T:      e.Ts,
@@ -713,4 +723,16 @@ func (a *appender) Rollback() error {
 	a.exemplars = a.exemplars[:0]
 	a.w.appenderPool.Put(a)
 	return nil
+}
+
+func exemplarsEqual(me memExemplar, pe exemplar.Exemplar) bool {
+	if !labels.Equal(me.labels, pe.Labels) {
+		return false
+	}
+
+	if me.ts != pe.Ts {
+		return false
+	}
+
+	return me.value == pe.Value
 }