--- a/pkg/metrics/wal/wal_test.go
+++ b/pkg/metrics/wal/wal_test.go
@@ -380,9 +380,9 @@ func BenchmarkAppendExemplar(b *testing.B) {
 	defer s.Close()
 	app := s.Appender(context.Background())
 	sRef, _ := app.Append(0, labels.Labels{{Name: "a", Value: "1"}}, 0, 0)
+	e := exemplar.Exemplar{Labels: labels.Labels{{Name: "a", Value: "1"}}, Value: 20, Ts: 10, HasTs: true}
 
 	b.StartTimer()
-	e := exemplar.Exemplar{Labels: labels.Labels{{Name: "a", Value: "1"}}, Value: 20, Ts: 10, HasTs: true}
 	for i := 0; i < b.N; i++ {
 		e.Ts = int64(i)
 		_, _ = app.AppendExemplar(sRef, nil, e)