#### PR Description 
This PR attempts to stop the Grafana Agent from sending duplicate exemplars, to avoid incurring extra I/O costs to the receiving side.

To achieve that, it adds a memExemplar struct to keep exemplars in memory. It's used to cache the latest exemplar (for series that have them) in a new field of stripeSeries, utilizing the same mechanism to distribute the lock contention.

#### Which issue(s) this PR fixes 
This fixes #1282 

#### Notes to the Reviewer
- I sketched out this with a couple of other approaches, either moving this one level up as a separate entry in the `Storage` struct, or in a separate `exemplars.go` file next to `series.go`. I didn't really like either of those options, I feel that this might be an acceptable solution comparing the added complexity with the benefit of keeping the code in one place
- Resources-wise, this will allocate defaultStripeSize pointers on startup (~130KB). Each series that has at least one exemplar appended will incur some move overhead (95 bytes on avg, 328 bytes on worst case) as an extra heap allocation.
- Running a benchmark that includes appending a valid exemplar (with its timestamp set to b.N in the benchmark loop) to gauge the performance of `AppendExemplar` before and after this change gives us

```
name              old time/op    new time/op    delta
AppendExemplar-8     255ns ±21%     332ns ±16%  +30.32%  (p=0.000 n=10+10)

name              old alloc/op   new alloc/op   delta
AppendExemplar-8      805B ± 4%      872B ± 7%   +8.37%  (p=0.001 n=10+10)

name              old allocs/op  new allocs/op  delta
AppendExemplar-8      2.00 ± 0%      3.00 ± 0%  +50.00%  (p=0.000 n=10+10)
```


#### PR Checklist

- [x] CHANGELOG updated 
- [x] Documentation added (N/A)
- [x] Tests updated