@rfratto 2022-01-26
> Running a benchmark that includes appending a valid exemplar (with its timestamp set to b.N in the benchmark loop) to gauge the performance of AppendExemplar before and after this change 

Love the benchmark stats! We should consider including the benchmark you wrote in the PR so we can run it again for future changes.