@tpaschalis 2022-01-26
> One extra thing: we should consider pre-populating `stripeSeries.exemplars` [like we do with `stripeSeries.series`](https://github.com/grafana/agent/blob/459b277c09abe17958ac020b4eb1ef7eff27322f/pkg/metrics/wal/wal.go#L303-L307) on initial startup of the WAL storage.
> 
> If we don't do this, then we might duplicate exemplars once per initialization of the WAL. It's likely not a huge deal, but would be consistent with how we handle other data in the WAL.

@rfratto Right now, we don't in decode exemplars in `loadWAL`. Should we also handle them during initialization in the same way there?

https://github.com/grafana/agent/blob/459b277c09abe17958ac020b4eb1ef7eff27322f/pkg/metrics/wal/wal.go#L283-L285