@rfratto 2022-01-26
> @rfratto Right now, we don't in decode exemplars in loadWAL. Should we also start handling them during initialization there?

Yeah, we'd need to start decoding them if we were going to do this :)