@cstyan 2022-01-26
:+1: +1 for importing the prometheus exemplar code and reusing it's `Equals` function.